+++ src/Phan/Parse/ParseVisitor.php
@@ -29,11 +29,14 @@ use Phan\Language\Type;
 use Phan\Language\Type\ArrayShapeType;
 use Phan\Language\Type\ArrayType;
 use Phan\Language\Type\CallableType;
+use Phan\Language\Type\MixedType;
 use Phan\Language\Type\NullType;
 use Phan\Language\Type\StringType;
 use Phan\Language\UnionType;
 use Phan\Library\None;
 use ast\Node;
+use ast;
+use InvalidArgumentException;
 
 /**
  * The class is a visitor for AST nodes that does parsing. Each
@@ -580,13 +583,22 @@ class ParseVisitor extends ScopeVisitor
 
             $value_node = $child_node->children['value'];
             if ($value_node instanceof Node) {
-                $constant->setFutureUnionType(
-                    new FutureUnionType(
-                        $this->code_base,
-                        $this->context,
-                        $value_node
-                    )
-                );
+                try {
+                    self::checkIsAllowedInConstExpr($value_node);
+                    $constant->setFutureUnionType(
+                        new FutureUnionType(
+                            $this->code_base,
+                            $this->context,
+                            $value_node
+                        )
+                    );
+                } catch (InvalidArgumentException $e) {
+                    $constant->setUnionType(MixedType::instance(false)->asUnionType());
+                    $this->emitIssue(
+                        Issue::InvalidConstantExpression,
+                        $value_node->lineno
+                    );
+                }
             } else {
                 $constant->setUnionType(Type::fromObject($value_node)->asUnionType());
             }
@@ -616,10 +628,23 @@ class ParseVisitor extends ScopeVisitor
         foreach ($node->children as $child_node) {
             \assert($child_node instanceof Node);
 
+            $value_node = $child_node->children['value'];
+            try {
+                self::checkIsAllowedInConstExpr($value_node);
+            } catch (InvalidArgumentException $e) {
+                $this->emitIssue(
+                    Issue::InvalidConstantExpression,
+                    $value_node->lineno
+                );
+                // Note: Global constants with invalid value expressions aren't declared.
+                // However, class constants are declared with placeholders to make inheritance checks, etc. easier.
+                // Both will emit PhanInvalidConstantExpression
+                continue;
+            }
             $this->addConstant(
                 $child_node,
                 $child_node->children['name'],
-                $child_node->children['value'],
+                $value_node,
                 $child_node->flags ?? 0,
                 $child_node->children['docComment'] ?? ''
             );
@@ -1243,5 +1268,49 @@ class ParseVisitor extends ScopeVisitor
     public function visitBinaryOp(Node $node)
     {
         return $this->context;
+    }
+
+    /**
+     * @internal
+     */
+    const ALLOWED_CONST_EXPRESSION_KINDS = [
+        ast\AST_ARRAY_ELEM => true,
+        ast\AST_ARRAY => true,
+        ast\AST_BINARY_OP => true,
+        ast\AST_CLASS_CONST => true,
+        ast\AST_COALESCE => true,
+        ast\AST_CONDITIONAL => true,
+        ast\AST_CONST => true,
+        ast\AST_DIM => true,
+        ast\AST_MAGIC_CONST => true,
+        ast\AST_NAME => true,
+        ast\AST_UNARY_OP => true,
+    ];
+
+    /**
+     * This is meant to avoid causing errors in Phan where Phan expects a constant to be found.
+     *
+     * @param Node|string|float|int|bool|null $n
+     *
+     * @throws InvalidArgumentException if this is not allowed in a constant expression
+     * Based on zend_bool zend_is_allowed_in_const_expr from Zend/zend_compile.c
+     *
+     * @internal
+     */
+    public static function checkIsAllowedInConstExpr($n) {
+        if (!($n instanceof Node)) {
+            if (\is_array($n)) {
+                foreach ($n as $child_node) {
+                    self::checkIsAllowedInConstExpr($child_node);
+                }
+            }
+            return;
+        }
+        if (!\array_key_exists($n->kind, self::ALLOWED_CONST_EXPRESSION_KINDS)) {
+            throw new InvalidArgumentException();
+        }
+        foreach ($n->children as $child_node) {
+            self::checkIsAllowedInConstExpr($child_node);
+        }
     }
 }